+++ (0)routes/auth.routes.js
@@ -3,7 +3,7 @@ const express = require("express");
 const router = express.Router();
 
 const authMiddleware = require("../middlewares/auth-middleware");
-const UserController = require("../controllers/users.controller");
+const UserController = require("../(1)controllers/users.controller");
 const userController = new UserController();
