+++ (0)routes/comment.routes.js
@@ -1,7 +1,8 @@
 const express = require("express");
 const router = express.Router();
+const CommentsController = require("../(1)controllers/comments.controller");
+const commentsController = new CommentsController();
 const authMiddleware = require("../middlewares/auth-middleware");
-
 
 
 module.exports = router;
