+++ (0)routes/like.routes.js
@@ -1,8 +1,9 @@
 const express = require("express");
 const router = express.Router();
-const CommentsController = require("../controllers/comments.controller");
-const commentsController = new CommentsController();
 const authMiddleware = require("../middlewares/auth-middleware");
+
+const LikesController = require("../(1)controllers/likes.controller");
+const likesController = new LikesController();
 
 
 module.exports = router;
