+++ (0)routes/post.routes.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
 const authMiddleware = require("../middlewares/auth-middleware");
-const PostController = require("../controllers/posts.controller");
+const PostController = require("../(1)controllers/posts.controller");
 const postController = new PostController();
 const uploadImage = require("../modules/s3.js");
