+++ (1)controllers/likes.controller.js
@@ -1,5 +1,5 @@
-const LikesService = require("../services/likes.service");
-const PostsService = require("../services/posts.service");
+const LikesService = require("../(2)services/likes.service");
+const PostsService = require("../(2)services/posts.service");
 
 class LikesController {
   likesService = new LikesService();
