+++ (2)services/comment.service.js
@@ -1,4 +1,4 @@
-const CommentRepository = require("../repositories/comments.repository");
+const CommentRepository = require("../(3)repositories/comment.repository");
 const { Comments } = require("../models");
 
 class CommentService {
