+++ (2)services/likes.service.js
@@ -1,6 +1,6 @@
 const { Likes, Comments } = require("../models/index");
-const LikesRepository = require("../repositories/likes.repository");
-const CommentsRepository = require("../repositories/comments.repository");
+const LikesRepository = require("../(3)repositories/likes.repository");
+const CommentsRepository = require("../(3)repositories/comment.repository");
 const { sequelize } = require("../models");
 const { Transaction } = require("sequelize");
