+++ (2)services/mypage.service.js
@@ -1,4 +1,4 @@
-const mypageRepository = require("../repositories/mypage.repository");
+const mypageRepository = require("../(3)repositories/mypage.repository");
 const { UserInfos, Likes, Posts } = require("../models");
 class MypageService {
