+++ (2)services/posts.service.js
@@ -1,0 +1,8 @@
+const PostRepository = require("../(3)repositories/posts.repository");
+const { post } = require("../(0)routes");
+const postRepository = new PostRepository();
+
+class PostService {
+}
+
+module.exports = PostService;
