+++ middlewares/auth-middleware.js
@@ -1,6 +1,6 @@
 const jwt = require("jsonwebtoken");
 const { Users } = require("../models");
-const TokenRepository = require("../repositories/tokens.repository");
+const TokenRepository = require("../(3)repositories/tokens.repository");
 
 const tokenRepository = new TokenRepository();
