--- repositories/tokens.repository.js
@@ -1,7 +1,0 @@
-const { Tokens } = require("../models");
-
-class TokenRepository {
-
-}
-
-module.exports = TokenRepository;
