--- services/posts.service.js
@@ -1,8 +1,0 @@
-const PostRepository = require("../repositories/posts.repository");
-const { post } = require("../routes");
-const postRepository = new PostRepository();
-
-class PostService {
-}
-
-module.exports = PostService;
